+++ forge/src/main/java/com/dudko/blazinghot/multiloader/forge/MultiRegistriesImpl.java
@@ -35,4 +35,8 @@ public class MultiRegistriesImpl {
 	public static ResourceLocation getFluidId(Fluid fluid) {
 		return ForgeRegistries.FLUIDS.getKey(fluid);
 	}
+
+	public static ResourceLocation getItemId(Item item) {
+		return ForgeRegistries.ITEMS.getKey(item);
+	}
 }